Separate logic in another file: uiSourceSubfield
